Revert "[ML] Hyper-parameter tuning via Genetic Algorithm (#6713)"
This reverts commit 63fbcbf8

+++ modules/ml/src/main/java/org/apache/ignite/ml/preprocessing/normalization/NormalizationTrainer.java
@@ -54,9 +54,9 @@ public class NormalizationTrainer<K, V> implements PreprocessingTrainer<K, V> {
      * @param p The given value.
      * @return The Normalization trainer.
      */
-    public NormalizationTrainer<K, V> withP(double p) {
+    public NormalizationTrainer<K, V> withP(int p) {
         assert p > 0;
-        this.p = (int) p;
+        this.p = p;
         return this;
     }
 }
